Add files via upload
Problem

Given two strings ss and tt (of possibly different lengths), the edit distance dE(s,t)dE(s,t) is the minimum number of edit operations needed to transform ss into tt, where an edit operation is defined as the substitution, insertion, or deletion of a single symbol.

The latter two operations incorporate the case in which a contiguous interval is inserted into or deleted from a string; such an interval is called a gap. For the purposes of this problem, the insertion or deletion of a gap of length kk still counts as kk distinct edit operations.

Given: Two protein strings ss and tt in FASTA format (each of length at most 1000 aa).

Return: The edit distance dE(s,t)dE(s,t).